+++ duckscript/src/lib_test.rs
@@ -1,0 +1,8 @@
+use super::*;
+
+#[test]
+fn version_test() {
+    let version = version();
+
+    assert!(!version.is_empty());
+}