+++ duckscript_sdk/src/sdk/std/debug/duckscript_version/mod.rs
@@ -1,0 +1,35 @@
+use crate::utils::pckg;
+use duckscript;
+use duckscript::types::command::{Command, CommandResult};
+
+#[cfg(test)]
+#[path = "./mod_test.rs"]
+mod mod_test;
+
+struct CommandImpl {
+    package: String,
+}
+
+impl Command for CommandImpl {
+    fn name(&self) -> String {
+        pckg::concat(&self.package, "DuckscriptVersion")
+    }
+
+    fn aliases(&self) -> Vec<String> {
+        vec!["duckscript_version".to_string()]
+    }
+
+    fn help(&self) -> String {
+        include_str!("help.md").to_string()
+    }
+
+    fn run(&self, _arguments: Vec<String>) -> CommandResult {
+        CommandResult::Continue(Some(duckscript::version()))
+    }
+}
+
+pub(crate) fn create(package: &str) -> Box<dyn Command> {
+    Box::new(CommandImpl {
+        package: package.to_string(),
+    })
+}